+++ metadetect/lsst/measure.py
@@ -235,7 +235,7 @@ def measure(
 
         if flags != 0:
             this_res = get_wavg_output_struct(nband=nband, model=fitter.kind)
-            this_res['flags'] = flags
+            this_res[fitter.kind + '_flags'] = flags
 
         res = get_output(
             wcs=wcs, source=source, res=this_res,