+++ CHANGELOG.md
@@ -4,9 +4,13 @@
 
 ### added
 
+ - Added `fwhm_reg` parameter to regularize the moments.
+
 ### changed
 
 ### removed
+
+ - Rolled back bugfix in 0.8.1 for the PSFs and `fwhm_smooth`.
 
 ### fixed