+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
  - Added `fwhm_reg` parameter to regularize the moments.
  - Added support for more than one measurement made on the detections.
+ - Added total image mask fraction to outputs.
 
 ### changed